Add the battle component and made it get the selected hero from draft
