--- src/components/menu/Battle.jsx
@@ -1,3 +1,0 @@
-export default function Battle() {
-  return <>Battle</>;
-}
