--- src/components/menu/Deck.jsx
@@ -1,5 +1,0 @@
-import Cards from '../Cards/Cards';
-
-export default function Deck() {
-  return <Cards />;
-}
